FIX CREATE MAKEFILE (Developpement server):
description :
Major change in this commit : A makefile is inside the front folder.
The dev server will not make a autonomous build and running,
reason: we work at the same time it compile.
So a makefile is needed for handling the server correctly
Production Version will create a folder dist after the build.

Features available in this version (stable):
    * make all : start the container (prompt will give help, read it :-))
    * when app is not installed : make install (all target will request you install target)
    * help : interactive guide after install, this is make all feature
    * make linter : make sure the project is according to our norm (make install required)

Feature available (no critical - unstable - fix soon):
    * make check_norm_bypass : this target create a report when developper think norm is broke
	fix : need to sequence each step inside target dependency, then abort script when needed

Next Feature before PR:
    * folder restructuration inside ./app/src (wiki page creation with tree view)