ADD CREATED ROOM IN THE GAME:
There are now two players per room.
Each time a new player is connected, a new room is created.
You can created an infinite number of rooms, we do not limit the server (for now)
In the same way, we don't reattribute the rooms that were left to new users.
Simple as that.